pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "open-sentinel"
version = "0.1.0"
description = "Reliability layer for AI agents - monitors workflow adherence and intervenes when agents deviate"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Open Sentinel Team" }
]
keywords = ["ai", "agents", "reliability", "workflow", "monitoring", "llm", "proxy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "litellm[proxy]>=1.50.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "httpx>=0.25.0",
]
examples = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "httpx>=0.25.0",
]
nemo = [
    "nemoguardrails>=0.10.0",
]
all = [
    "nemoguardrails>=0.10.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "httpx>=0.25.0",
]

[project.scripts]
osentinel = "opensentinel.cli:main"

[project.urls]
Homepage = "https://github.com/open-sentinel/open-sentinel"
Documentation = "https://github.com/open-sentinel/open-sentinel#readme"
Repository = "https://github.com/open-sentinel/open-sentinel"

[tool.hatch.build.targets.wheel]
packages = ["opensentinel"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
